requirements.txt
altair==5.4.1
annotated-types==0.7.0   
anyio==4.6.2.post1       
attrs==24.2.0
autogen==0.3.2
beautifulsoup4==4.12.3   
blinker==1.9.0
cachetools==5.5.0        
certifi==2024.8.30       
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
FLAML==2.3.2
fonttools==4.55.0        
frozendict==2.4.6        
gitdb==4.0.11
GitPython==3.1.43        
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
jiter==0.7.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
mplfinance==0.12.10b0
multitasking==0.0.11
narwhals==1.13.5
numpy==1.26.4
openai==1.54.4
packaging==24.2
pandas==2.2.3
peewee==3.17.8
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
protobuf==5.28.3
pyarrow==18.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
streamlit==1.40.1
tenacity==9.0.0
termcolor==2.5.0
tiktoken==0.8.0
toml==0.10.2
tornado==6.4.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
webencodings==0.5.1
yfinance==0.2.49